Rename Params To CallInspector and move it to packages folder
fixes #127
