+++ src/hook/useCouncilMembers.ts
@@ -12,15 +12,19 @@ interface Result {
 export function useCouncilMembers() {
   const {api} = useContext(ChainApiContext);
   const {accounts} = useAccounts();
-  return useQuery(['council-members', {accounts}], async () => {
-    const accountIds = await api?.query.council.members();
+  return useQuery(
+    ['council-members', {accounts}],
+    async () => {
+      const accountIds = await api?.query.council.members();
 
-    if (accountIds && api) {
-      const members = await getAccountsIdentityInfo(accountIds, api);
-      return {
-        members,
-        isMember: members.some((accountId) => accounts?.find((a) => a.address.toString() === accountId.toString())),
-      };
-    }
-  });
+      if (accountIds && api) {
+        const members = await getAccountsIdentityInfo(accountIds, api);
+        return {
+          members,
+          isMember: members.some((accountId) => accounts?.find((a) => a.address.toString() === accountId.toString())),
+        };
+      }
+    },
+    {staleTime: 0},
+  );
 }
